updates to Handlebar Templates
I discovered that I had overlooked data already in the JSON that could help derive the distance scale of the map for leaflet and with that info, it greatly reduced the complexity of adjusting x & y values for markers in Leaflet.

Also added a link to the FMG map stored in a Dropbox link. It requires adding that to the JSON file, but it allows links to be created that can take you directly to State/Province/Burg locations.