updates to Handlebars to use Dataview blocks for children
The lists on the individual State and Province notes now will use Dataview blocks to show the lists of Neighbors, Provinces, and Burgs within each location. Reducing hard-coded links in the Infobox sections.

Also removed the "[[ and ]]" that were put around states, provinces, and neighbors in frontmatter. Using the Dataview `link()` function means we can link without coding them as links in the YAML