small update to how State Id 0 is handled
State Id 0 was being skipped - this is usually the Neutrals or Wild Lands of an FMG world. I thought that should get at least a State Note so you'd have a place to create any Neutral/Wild info for your world map.

Also, discovered how the {{setvar}} handlebar helper is supposed to work, so will start switching things over to make use of that so we aren't making multiple calls to the same functions in an individual note.