Update Helpers-FMG-JSON.js
hiding the name text on the Village Gen burgMaps - text element is set to a very large size by the generator. FMG hides it by default & I thought it was OK to keep, but it's just too large and obscures much of the village map when viewed in a small window.